+++ HISTORY.md
@@ -3,6 +3,8 @@
 
 ## not yet released, version 1.1.1
 
+- Fixed `setTimeout` and `setTrigger` not working correctly when adding
+  dates in the past or infinite delays. 
 - Implemented a workaround for a bug in node.js when calling `setTimeout`
   with a non-integer delay.